Added solution - LeetHub

+++ Implement strstr - GFG/implement-strstr.cpp
@@ -26,7 +26,7 @@ int main()
 //Function to locate the occurrence of the string x in the string s.
 int strstr(string s, string x)
 {
-    size_t found=s.find(x);
+    int found=s.find(x);
     if (found != string::npos){
         return found;
     }
